resume-factory/requirements.txt
jinja2
pyyaml
pydantic
pylatex